+++ test-data/woo-products.wxr
@@ -2051,7 +2051,6 @@
   <wp:comment_status>open</wp:comment_status>
   <wp:ping_status>closed</wp:ping_status>
   <wp:post_name>album</wp:post_name>
-  <wp:status>publish</wp:status>
   <wp:post_parent>0</wp:post_parent>
   <wp:menu_order>0</wp:menu_order>
   <wp:post_type>product</wp:post_type>
